32-fds_enable_agent/requirements.txt
# Deep Learning
torch>=2.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# OpenAI and LLM
openai>=1.12.0

# Data serialization
pyyaml>=6.0
joblib>=1.3.0

# Environment variables
python-dotenv>=1.0.0
